+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ### Changed
 - Changed agent namespace to default to the pod namespace.
   Closes [[#404](https://github.com/metalbear-co/mirrord/issues/404)].
+- mirrord-layer: Use `tracing::instrument` to improve logs.
 
 ## 3.0.2-alpha